backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
requests
Pillow
feedparser
